// src/interfaces/color.interface.ts
export type SemanticColors<V = string> = Record<
    SemanticColorNames,
    V
>

export type SemanticColorNames
    = 'PrimaryKeyColor'
    | 'SecondaryKeyColor'
    | 'TertiaryKeyColor'
    | 'ErrorKeyColor'
    | 'NeutralKeyColor'
    | 'NeutralVariantKeyColor'
    | 'Background'
    | 'Error'
    | 'ErrorContainer'
    | 'ErrorDim'
    | 'InverseOnSurface'
    | 'InversePrimary'
    | 'InverseSurface'
    | 'OnBackground'
    | 'OnError'
    | 'OnErrorContainer'
    | 'OnPrimary'
    | 'OnPrimaryContainer'
    | 'OnPrimaryFixed'
    | 'OnPrimaryFixedVariant'
    | 'OnSecondary'
    | 'OnSecondaryContainer'
    | 'OnSecondaryFixed'
    | 'OnSecondaryFixedVariant'
    | 'OnSurface'
    | 'OnSurfaceVariant'
    | 'OnTertiary'
    | 'OnTertiaryContainer'
    | 'OnTertiaryFixed'
    | 'OnTertiaryFixedVariant'
    | 'Outline'
    | 'OutlineVariant'
    | 'Primary'
    | 'PrimaryContainer'
    | 'PrimaryDim'
    | 'PrimaryFixed'
    | 'PrimaryFixedDim'
    | 'Scrim'
    | 'Secondary'
    | 'SecondaryContainer'
    | 'SecondaryDim'
    | 'SecondaryFixed'
    | 'SecondaryFixedDim'
    | 'Shadow'
    | 'Surface'
    | 'SurfaceBright'
    | 'SurfaceContainer'
    | 'SurfaceContainerHigh'
    | 'SurfaceContainerHighest'
    | 'SurfaceContainerLow'
    | 'SurfaceContainerLowest'
    | 'SurfaceDim'
    | 'SurfaceTint'
    | 'SurfaceVariant'
    | 'Tertiary'
    | 'TertiaryContainer'
    | 'TertiaryDim'
    | 'TertiaryFixed'
    | 'TertiaryFixedDim'